+++ src/lib.rs
@@ -92,7 +92,7 @@ pub use nodemap::{
     Attributes, Entry, MutableAttributes, MutableNamespaces, MutableNodeMap, Namespaces, NodeMap,
 };
 pub use parse::{Span, SpanInfo, SpanInfoKey};
-pub use serialize::{PrettyOutputToken, SerializeOptions, WithSerializeOptions};
+pub use serialize::PrettyOutputToken;
 pub use serializer::{Output, OutputToken};
 pub use xmlvalue::{
     Attribute, Comment, Element, Namespace, Prefixes, ProcessingInstruction, Text, Value, ValueType,