+++ src/serialize.rs
@@ -25,45 +25,14 @@ pub struct PrettyOutputToken {
     pub newline: bool,
 }
 
-/// Options to control serialization
-#[derive(Debug, Default)]
-pub struct SerializeOptions {
-    /// Pretty print XML
-    pub pretty: bool,
-}
-
-/// Configurable serialization
-pub struct WithSerializeOptions<'a> {
-    xot: &'a Xot,
-    options: SerializeOptions,
-}
-
-impl<'a> WithSerializeOptions<'a> {
-    /// Write node as XML.
-    pub fn write(&self, node: Node, w: &mut impl Write) -> Result<(), Error> {
-        let outputs = gen_outputs(self.xot, node);
-        let mut serializer = XmlSerializer::new(self.xot, node);
-        if self.options.pretty {
-            serializer.serialize_pretty(w, outputs, vec![])
-        } else {
-            serializer.serialize(w, outputs)
-        }
-    }
-
-    /// Write node to XML string.
-    pub fn to_string(&self, node: Node) -> Result<String, Error> {
-        let mut buf = Vec::new();
-        self.write(node, &mut buf)?;
-        Ok(String::from_utf8(buf).unwrap())
-    }
-}
-
 /// ## Serialization
 impl Xot {
     /// Write node as XML.
     ///
-    /// You can control output options by using [`Xot::with_serialize_options`] first,
-    /// and calling `write` on that.
+    /// This uses the default serialization parameters: no XML declaration, no
+    /// doctype, no control over pretty printing or CDATA.
+    ///
+    /// To serialize with this control use [`Xot::serialize_xml_write`].
     ///
     /// If there are missing namespace prefixes, this errors. You can
     /// automatically add missing prefixes by invoking
@@ -83,14 +52,17 @@ impl Xot {
     /// # Ok::<(), xot::Error>(())
     /// ```
     pub fn write(&self, node: Node, w: &mut impl Write) -> Result<(), Error> {
-        self.with_serialize_options(SerializeOptions::default())
-            .write(node, w)
+        let outputs = gen_outputs(self, node);
+        let mut serializer = XmlSerializer::new(self, node, vec![]);
+        serializer.serialize(w, outputs)
     }
 
     /// Serialize node as XML string.
     ///
-    /// You can control output options by using [`Xot::with_serialize_options`] first,
-    /// and calling `to_string` on that.
+    /// This uses the default serialization parameters: no XML declaration, no
+    /// doctype, no control over pretty printing or CDATA.
+    ///
+    /// To serialize with this control use [`Xot::serialize_xml_string`].
     ///
     /// If there are missing namespace prefixes, this errors. You can automatically
     /// add missing prefixes by invoking [`Xot::create_missing_prefixes`] before
@@ -108,11 +80,19 @@ impl Xot {
     /// # Ok::<(), xot::Error>(())
     /// ```
     pub fn to_string(&self, node: Node) -> Result<String, Error> {
-        self.with_serialize_options(SerializeOptions::default())
-            .to_string(node)
+        let mut buf = Vec::new();
+        self.write(node, &mut buf)?;
+        Ok(String::from_utf8(buf).unwrap())
     }
 
     /// Serialize to XML, with options.
+    ///
+    /// Note that if you don't need string output and have a writer available,
+    /// a more efficient option is to use [`Xot::serialize_xml_write`].
+    ///
+    /// If there are missing namespace prefixes, this errors. You can automatically
+    /// add missing prefixes by invoking [`Xot::create_missing_prefixes`] before
+    /// serialization to avoid this error.
     ///
     /// With the default parameters:
     /// ```rust
@@ -121,7 +101,7 @@ impl Xot {
     /// let mut xot = Xot::new();
     /// let root = xot.parse("<a><b/></a>")?;
     ///
-    /// let xml = xot.serialize_xml(Default::default(), root)?;
+    /// let xml = xot.serialize_xml_string(Default::default(), root)?;
     /// assert_eq!(xml, "<a><b/></a>");
     /// # Ok::<(), xot::Error>(())
     /// ```
@@ -134,7 +114,7 @@ impl Xot {
     /// let mut xot = Xot::new();
     /// let root = xot.parse("<a><b/></a>")?;
     ///
-    /// let xml = xot.serialize_xml(output::xml::Parameters {
+    /// let xml = xot.serialize_xml_string(output::xml::Parameters {
     ///     declaration: Some(Default::default()),
     ///     ..Default::default()
     /// }, root)?;
@@ -142,7 +122,8 @@ impl Xot {
     /// # Ok::<(), xot::Error>(())
     /// ```
     ///
-    /// XML declaration with an encoding declaration (does not affect output encoding):
+    /// XML declaration with an encoding declaration (does not affect output
+    /// encoding):
     ///
     /// ```rust
     /// use xot::{Xot, output};
@@ -150,7 +131,7 @@ impl Xot {
     /// let mut xot = Xot::new();
     /// let root = xot.parse("<a><b/></a>")?;
     ///
-    /// let xml = xot.serialize_xml(output::xml::Parameters {
+    /// let xml = xot.serialize_xml_string(output::xml::Parameters {
     ///     declaration: Some(output::xml::Declaration {
     ///         encoding: Some("UTF-8".to_string()),
     ///         ..Default::default()
@@ -169,21 +150,36 @@ impl Xot {
     /// let mut xot = Xot::new();
     /// let root = xot.parse("<a><b/></a>")?;
     ///
-    /// let xml = xot.serialize_xml(output::xml::Parameters {
+    /// let xml = xot.serialize_xml_string(output::xml::Parameters {
     ///     indentation: Some(Default::default()),
     ///     ..Default::default()
     /// }, root)?;
     /// assert_eq!(xml, "<a>\n  <b/>\n</a>\n");
     /// # Ok::<(), xot::Error>(())
     /// ```
-    pub fn serialize_xml(
+    pub fn serialize_xml_string(
         &self,
         parameters: output::xml::Parameters,
         node: Node,
     ) -> Result<String, Error> {
         let mut buf = Vec::new();
+        self.serialize_xml_write(parameters, node, &mut buf)?;
+        Ok(String::from_utf8(buf).unwrap())
+    }
+
+    /// Serialize to XML via a [`Write`], with options.
+    ///
+    /// This is like [`Xot::serialize_xml_string`] but writes to a [`Write`]. This
+    /// is more efficient if you want to write directly to a file, for
+    /// instance, as no string needs to be created in memory.
+    pub fn serialize_xml_write(
+        &self,
+        parameters: output::xml::Parameters,
+        node: Node,
+        w: &mut impl Write,
+    ) -> Result<(), Error> {
         if let Some(declaration) = parameters.declaration {
-            declaration.serialize(&mut buf)?;
+            declaration.serialize(w)?;
         }
         if let Some(doctype) = parameters.doctype {
             // if we are in a document node, we look for the document_element,
@@ -197,36 +193,17 @@ impl Xot {
             // know it's an element now
             let name = self.node_name_ref(node)?.unwrap();
             let name = name.full_name();
-            doctype.serialize(name.as_ref(), &mut buf)?;
+            doctype.serialize(name.as_ref(), w)?;
         }
         let outputs = gen_outputs(self, node);
-        let mut serializer = XmlSerializer::new(self, node);
+        let mut serializer =
+            XmlSerializer::new(self, node, parameters.cdata_section_elements.clone());
         if let Some(indentation) = parameters.indentation {
-            serializer.serialize_pretty(&mut buf, outputs, indentation.suppress.clone())?;
+            serializer.serialize_pretty(w, outputs, indentation.suppress.clone())?;
         } else {
-            serializer.serialize(&mut buf, outputs)?;
+            serializer.serialize(w, outputs)?;
         }
-        Ok(String::from_utf8(buf).unwrap())
-    }
-
-    /// Control XML serialization
-    ///
-    /// You can control the serialization before invoking [`WithSerializeOptions::write`] or
-    /// [`WithSerializeOptions::to_string`] by passing in options.
-    ///
-    /// ```rust
-    /// use xot::{Xot, SerializeOptions};
-    ///
-    /// let mut xot = Xot::new();
-    /// let root = xot.parse("<a><b/></a>")?;
-    ///
-    /// let buf = xot.with_serialize_options(SerializeOptions { pretty: true, ..SerializeOptions::default() }).to_string(root)?;
-    ///
-    /// assert_eq!(buf, "<a>\n  <b/>\n</a>\n");
-    /// # Ok::<(), xot::Error>(())
-    /// ```
-    pub fn with_serialize_options(&self, options: SerializeOptions) -> WithSerializeOptions {
-        WithSerializeOptions { xot: self, options }
+        Ok(())
     }
 
     /// Serialize node into outputs.
@@ -255,7 +232,7 @@ impl Xot {
     /// a different way, for instance with inline styling.
     pub fn tokens(&self, node: Node) -> impl Iterator<Item = (Node, Output, OutputToken)> + '_ {
         let outputs = gen_outputs(self, node);
-        let mut serializer = XmlSerializer::new(self, node);
+        let mut serializer = XmlSerializer::new(self, node, vec![]);
         outputs.map(move |(node, output)| {
             let rendered = serializer.render_output(node, &output).unwrap();
             (node, output, rendered)
@@ -272,7 +249,7 @@ impl Xot {
         node: Node,
     ) -> impl Iterator<Item = (Node, Output, PrettyOutputToken)> + '_ {
         let outputs = gen_outputs(self, node);
-        let mut serializer = XmlSerializer::new(self, node);
+        let mut serializer = XmlSerializer::new(self, node, vec![]);
         let mut pretty = Pretty::new(self, vec![]);
         outputs.map(move |(node, output)| {
             let (indentation, newline) = pretty.prettify(node, &output);